refactor: partition object import/export modules
